+++ src/main/java/io/github/zatone0/ae2craftingrecovery/mixin/ExecutingCraftingJobAccessor.java
@@ -8,6 +8,7 @@ import org.spongepowered.asm.mixin.gen.Accessor;
 import appeng.api.crafting.IPatternDetails;
 import appeng.api.stacks.GenericStack;
 import appeng.crafting.CraftingLink;
+import appeng.crafting.execution.ElapsedTimeTracker;
 import appeng.crafting.execution.ExecutingCraftingJob;
 import appeng.crafting.inv.ListCraftingInventory;
 
@@ -27,6 +28,9 @@ public interface ExecutingCraftingJobAccessor {
 
     @Accessor("remainingAmount")
     long ae2cr$getRemainingAmount();
+
+    @Accessor("timeTracker")
+    ElapsedTimeTracker ae2cr$getTimeTracker();
 
     @Accessor("playerId")
     Integer ae2cr$getPlayerId();